DIAF grouper + n1
Grouper is very broken on about 7 levels and not a single fix me request in months

N1 I am the only dev on github to make a M6 commit since Novemeber.  Testers do NOT follow up
If you want this to work send me freaking logs! Or we stop maintaining it.

Change-Id: I3fdd1af798deb6eba360e8d83358bc73a1024ff1